Fix issue with transaction detection
Prior to this change we were filtering the transaction lists and only
looking at 'charges'. I incorrectly determined that 'charges' were
payments to me. This is actually not the case, a 'charge' is simply when
I _request_ payment, and a `pay` is when someone else _pays me_.

We now filter the transaction list by specifically looking if the payee
is me, ignoring the transaction_type completely.